The UserScripts control logic has been moved to lib-iitc-manager. This library is covered by tests and some bugs are fixed.
