+++ src/popup/components/mixins.js
@@ -1,5 +1,5 @@
 //@license magnet:?xt=urn:btih:1f739d935676111cfff4b4693e3816e664797050&dn=gpl-3.0.txt GPL-v3
-import { _ } from "../../helpers.js";
+import { _ } from "@/i18n";
 
 export const mixin = {
   data() {
